refactor: address code review comments
